Refactor: 구조 관리
파일 분리하기
