+++ src/App.js
@@ -1,11 +1,12 @@
 import {BrowserRouter as Router, Routes, Route} from "react-router-dom";
-import "./App.css";
-import Header from "./components/Header";
+
+import Header from "./components/common/Header";
 import Main from "./page/Main";
 import Train from "./page/Train"
 import Map from "./page/Map";
 import Bus from "./page/Bus";
 import TrainResult from "./page/TrainResult";
+import "./App.css";
 
 function App() {
   return (
@@ -18,7 +19,6 @@ function App() {
           <Route exact path="/map" element={<Map />}/>
           <Route exact path="/train-result" element={<TrainResult />} />
           <Route exact path="/bus/:date" element={<Bus />}/>
-
         </Routes>
       </Router>
     </div>
