+++ src/components/slide/Slide.css
