+++ src/components/slide/Slide.js
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react'
 import { Link } from 'react-router-dom';
-import Button from './Button'
+import Button from '../common/Button'
 import "./Slide.css";
 const IMAGE_LENGTH = 3;
