+++ src/page/Main.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Slide from '../components/Slide'
+import Slide from '../components/slide/Slide'
 import FooterMain from '../components/FooterMain';
 import "./Main.css";
