+++ src/page/TrainResult.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom';
 
-import Button from "../components/Button";
+import Button from "../components/common/Button";
 import { callData } from '../components/useData';
 import { TIME_TABLE_URL } from '../API/API_URL';
